refactor (#470)
